requirements.txt
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
idna==3.4
numpy==1.24.4
pandas==2.0.3
pur==7.1.0
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
requests==2.31.0
six==1.16.0
tqdm==4.65.0
tzdata==2023.3
urllib3==2.0.3
websocket-client==1.6.1